Rename home files
