+++ simplq/src/components/pages/Home/Home.module.scss
