+++ simplq/src/components/pages/Home/HomePage.jsx
