+++ simplq/src/components/pages/Home/StaticInfos.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import StandardButton from 'components/common/Button';
 import { smoothScrollToHomePageTop } from 'utils/scrollingOperations';
-import styles from './home.module.scss';
+import styles from './Home.module.scss';
 
 export const BenefitsInfo = () => (
   <div className={`${styles['section']} ${styles['benefits-info']}`}>
